Backend

See merge request s12-fintech-finance-sub1/S12P21D204!54

+++ backend/LocketPaymentService/src/main/java/com/locket/payment/domain/pay/entity/PaymentTransaction.java
@@ -26,10 +26,10 @@ public class PaymentTransaction {
     private BankAccount account;
 
     @Column(nullable = false)
-    private Integer buyerId;
+    private Long buyerId;
 
     @Column(nullable = false)
-    private Integer sellerId;
+    private Long sellerId;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
